rename package and docstrings
