+++ dns-server/dns-records.txt
@@ -1,0 +1,14 @@
+; This file contains a list of DNS records that will be loaded into our
+; DNS server's local store
+;
+acme.com.                       127.0.0.1
+blog.acme.com.                  127.0.0.1
+
+; Blocked domains
+; Domains in this list are blocked to prevent tracking of users' activity
+;
+mask.icloud.com.                BLOCK
+ads.google.com.                 BLOCK
+adservice.google.com.           BLOCK
+www.googletagmanager.com.       BLOCK
+cdn.googletoolservices.com.     BLOCK
