+++ dns-server/go.mod
@@ -1,0 +1,3 @@
+module github.com/mcastellin/golang-mastery/dns-server
+
+go 1.21
