--- dns/go.mod
@@ -1,3 +1,0 @@
-module github.com/mcastellin/golang-mastery/dns
-
-go 1.21
